refactor(cockpit): extract _SettingsConfirmModal into sibling module (#1354)
Third wedge of the cockpit_ui.py god-module split (after #1606 and
#1618). The generic confirm/cancel modal used by cockpit settings
screens is the cleanest remaining target — it was defined identically
in both cockpit_ui.py and cockpit_project_settings.py, so this
extraction also deduplicates ~57 LOC.

Moves the 55-LOC _SettingsConfirmModal into a new sibling module
src/pollypm/cockpit_settings_confirm.py. cockpit_ui.py keeps an import
shim for back-compat, and cockpit_project_settings.py imports from the
new module (dropping its now-unused textual.screen.ModalScreen import).

cockpit_ui.py: 18818 -> 18764 LOC (-54); diff in cockpit_ui.py is 60.
cockpit_project_settings.py: 1013 -> 956 LOC (-57).
